+++ src/html/default-icons.html
@@ -5,7 +5,7 @@
     <link type="text/css" rel="stylesheet" href="/lib/css/material-icons.css" />
     <link type="text/css" rel="stylesheet" href="/lib/css/materialize.min.css" media="screen,projection" />
     <link type="text/css" rel="stylesheet" href="/lib/css/roundslider.min.css" />
-    <link type="text/css" rel="stylesheet" href="/css/theme-editor.css" />
+    <link type="text/css" rel="stylesheet" href="/styles/theme-editor.css" />
     <link rel="shortcut icon" type="image/png" href="/imgs/icon@128.png" />
 
     <style>
