+++ src/styles/all.css
@@ -59,6 +59,10 @@ nav.nav-breadcrumb-nav-1251684582 span {
     color: var(--contrast-text-color) !important;
 }
 
+#splus-settings-navbar-button svg {
+    display: block;
+}
+
 /* Borders at the top of drop down cards (Courses and Groups) */
 
 ._2P7CY:focus,
@@ -135,6 +139,11 @@ footer._2T2dA {
     background-position: center center !important;
     height: 60px !important;
     margin-top: 0 !important;
+}
+
+/* March 2024 change to how custom logos are displayed */
+.use-custom-url div[data-sgy-sitenav="custom-header-logo"] {
+    visibility: hidden !important;
 }
 
 /* links inside the container for the added "Archived Courses" link within the Courses dropdown */
